fix: artist_details not loading because image height/width is returned as a float instead of a integer

+++ src/api/api_models.rs
@@ -1,6 +1,7 @@
 use form_urlencoded::Serializer;
 use regex::Regex;
 use serde::{Deserialize, Serialize};
+use serde_json::Number;
 use std::{
     collections::HashSet,
     convert::{Into, TryFrom, TryInto},
@@ -178,7 +179,8 @@ trait WithImages {
     }
 
     fn best_image_for_width(&self, width: i32) -> Option<&Image> {
-        self.best_image(|i| (width - i.width.unwrap_or(0) as i32).abs())
+        // When the API does return a float, it seems to just be an integer with a .0
+        self.best_image(|i| (width - (i.width.unwrap_or(0.0).abs()) as i32))
     }
 }
 
@@ -265,8 +267,9 @@ impl WithImages for Album {
 #[derive(Deserialize, Debug, Clone)]
 pub struct Image {
     pub url: String,
-    pub height: Option<u32>,
-    pub width: Option<u32>,
+    // Spotify's API sometimes returns image height/width as a float instead of a integer
+    pub height: Option<f64>,
+    pub width: Option<f64>,
 }
 
 #[derive(Deserialize, Debug, Clone)]
